feat(gpv2): organize and export types
