+++ src/entities/trades/gnosis-protocol/index.ts
@@ -1,1 +1,2 @@
 export * from './GnosisProtocolTrade'
+export * from './types'
